add square with default impl to IsField
