+++ math/src/field/extensions/quadratic.rs
@@ -60,6 +60,14 @@ where
         [&a0b0 + &a1b1 * q, z - a0b0 - a1b1]
     }
 
+    fn square(a: &[FieldElement<Q::BaseField>; 2]) -> [FieldElement<Q::BaseField>; 2] {
+        let [a0, a1] = a;
+        let v0 = a0 * a1;
+        let c0 = (a0 + a1) * (a0 + Q::residue() * a1) - &v0 - Q::residue() * &v0;
+        let c1 = &v0 + &v0;
+        [c0, c1]
+    }
+
     /// Returns the component wise subtraction of `a` and `b`
     fn sub(
         a: &[FieldElement<Q::BaseField>; 2],
@@ -118,13 +126,6 @@ where
 }
 
 impl<Q: Clone + Debug + HasQuadraticNonResidue> FieldElement<QuadraticExtensionField<Q>> {
-    pub fn square(&self) -> Self {
-        let [a0, a1] = self.value();
-        let v0 = a0 * a1;
-        let c0 = (a0 + a1) * (a0 + Q::residue() * a1) - &v0 - Q::residue() * &v0;
-        let c1 = &v0 + &v0;
-        Self::new([c0, c1])
-    }
 }
 
 #[cfg(test)]
